chore: rename runtime config'
