+++ storybook-static/index.html
@@ -56,4 +56,4 @@
           
       
           
-            window['DOCS_MODE'] = false;</script><script src="runtime~main.30fd60e3d03493dd6424.manager.bundle.js"></script><script src="878.faf19ee0e09ee8e8cca3.manager.bundle.js"></script><script src="main.be86fe191119f2eeec41.manager.bundle.js"></script></body></html>
+            window['DOCS_MODE'] = false;</script><script src="runtime~main.718aefac1aab85de4a24.manager.bundle.js"></script><script src="730.80e3228c3eb696183a89.manager.bundle.js"></script><script src="main.830751727662f2fc1ee1.manager.bundle.js"></script></body></html>
